+++ leo_fw/CHANGELOG.rst
@@ -2,6 +2,16 @@
 Changelog for package leo_fw
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+2.6.0 (2026-05-21)
+------------------
+* feat: Use EventsExecutor to reduce CPU consumption (`#44 <https://github.com/LeoRover/leo_robot-ros2/issues/44>`_)
+* feat: New firmware parameter bridging and node parameter override mechanism (`#43 <https://github.com/LeoRover/leo_robot-ros2/issues/43>`_)
+* fix: Ensure firmware parameters are always updated (`#41 <https://github.com/LeoRover/leo_robot-ros2/issues/41>`_)
+* Contributors: Błażej Sowa, Jan Hernas
+
+2.5.1 (2026-01-09)
+------------------
+
 2.5.0 (2025-12-29)
 ------------------
 * Use sphinx format docstrings (`#36 <https://github.com/LeoRover/leo_robot-ros2/issues/36>`_)